refact: BookServiceImpl 제거
